backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
asyncpg==0.29.0
databases==0.9.0
redis==5.0.4
upstash-redis==1.1.0
python-dotenv==1.0.1
pydantic==2.7.1
uuid==1.30
httpx==0.27.0